+++ internal/impl/mssqlserver/replication/stream_message.go
